+++ src/settings/WaiveSettings.js
@@ -45,7 +45,7 @@ class WaiveSettings extends React.Component {
         }}
         nameKey="waiveReasons"
         hiddenFields={['numberOfObjects']}
-        id="waives"
+        id="settings-waives"
         sortby="nameReason"
       />
     );